# swagger.yaml
openapi: 3.0.0
info:
  description: App that allows to transcribe speech to text
  version: 1.0.0
  title: Dicent
  contact:
    email: hello@neulen.dev
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
...
paths:
...
  "/users/{id}/projects":
    get:
      tags:
        - users
      summary: Get user projects
      description: ""
      operationId: getUserProjects
      parameters:
        - name: id
          in: path
          description: The id that needs to be fetched.
          required: true
          schema:
            type: integer
        - in: query
          name: ordDate
          description: Sort by date
          schema:
            type: string
            enum:
              - ASC
              - DESC
        - in: query
          name: search
          description: Filter by title
          schema:
            type: string
        - in: query
          name: status
          description: Filter by status
          schema:
            type: string
            enum:
              - NONE
              - QUEUED
              - IN_PROGRESS
              - FAILED
              - COMPLETED
        - in: query
          name: group
          description: Filter by group id
          schema:
            type: integer
        - in: query
          name: speakersFrom
          description: Filter by speakers num from
          schema:
            type: integer
        - in: query
          name: speakersTo
          description: Filter by speakers num to
          schema:
            type: integer
        - in: query
          name: page
          description: Page number
          schema:
            type: integer
        - in: query
          name: perPage
          description: Items on the page
          schema:
            type: integer
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ProjectDemo"
        "401":
          description: Permission denied
        "404":
          description: User not found
...
  "/groups":
    get:
      tags:
        - groups
      summary: Get user groups
      description: Only for authorized users.
      operationId: getGroups
      parameters:
        - in: query
          name: user_id
          description: Attach to user_id
          schema:
            type: integer
      responses:
        "200":
          description: Successfully fetched
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      tags:
        - groups
      summary: Create user group
      description: Only for authorized users.
      operationId: addGroup
      parameters:
        - in: query
          name: user_id
          description: Attach to user_id
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Group"
        description: Project object that needs to be added to the user
        required: true
      responses:
        "200":
          description: Successfully created
...
servers:
  - url: http://127.0.0.1/api
  - url: http://127.0.0.1
components:
  schemas:
    Group:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
...
    ProjectDemo:
      type: object
      required:
        - title
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        group:
          type: string
        speakers:
          type: integer
        job:
          type: string
        status:
          enum:
            - NONE
            - QUEUED
            - IN_PROGRESS
            - FAILED
            - COMPLETED
        audiofile:
          type: string
          description: Link to the audiofile